requirements.txt
# Core numerical and scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Jupyter environment
jupyter>=1.0.0
notebook>=7.0.0
ipython>=8.0.0

# Visualization
matplotlib>=3.7.0

# Date handling
python-dateutil>=2.8.0

# Statistical modeling
scikit-learn>=1.3.0

# Development tools
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Code linting
mypy>=1.5.0   # Type checking